+++ pytsite/auth_ui/res/assets/js/widget/follow.js
@@ -1,12 +1,15 @@
 $(function () {
     var widget = $('.widget-auth-ui-follow');
+    var follow_msg_id = widget.data('followMsgId');
+    var unfollow_msg_id = widget.data('unfollowMsgId');
+    var following_msg_id = widget.data('followingMsgId');
 
     widget.mouseover(function () {
         var btn = $(this).find('.btn');
         if (btn.hasClass('following')) {
             btn.removeClass('btn-primary').addClass('btn-danger');
             btn.find('.icon').removeClass('fa-check').addClass('fa-remove');
-            btn.find('.text').text(t('pytsite.auth_ui@unfollow'));
+            btn.find('.text').text(t(unfollow_msg_id));
         }
     });
 
@@ -15,7 +18,7 @@ $(function () {
         if (btn.hasClass('following')) {
             btn.removeClass('btn-danger').addClass('btn-primary');
             btn.find('.icon').removeClass('fa-remove').addClass('fa-check');
-            btn.find('.text').text(t('pytsite.auth_ui@following'));
+            btn.find('.text').text(t(following_msg_id));
         }
     });
     
@@ -26,7 +29,7 @@ $(function () {
                 if (typeof data.status != 'undefined' && data.status === true) {
                     btn.removeClass('btn-danger').addClass('btn-default').removeClass('following').addClass('non-following');
                     btn.find('.icon').addClass('fa-plus');
-                    btn.find('.text').text(t('pytsite.auth_ui@follow'));
+                    btn.find('.text').text(t(follow_msg_id));
                 }
             });
         }
@@ -35,7 +38,7 @@ $(function () {
                 if (typeof data.status != 'undefined' && data.status === true) {
                     btn.removeClass('btn-default').addClass('btn-danger').removeClass('non-following').addClass('following');
                     btn.find('.icon').removeClass('fa-plus').addClass('fa-remove');
-                    btn.find('.text').text(t('pytsite.auth_ui@unfollow'));
+                    btn.find('.text').text(t(unfollow_msg_id));
                 }
             });
         }