+++ pytsite/flag/__init__.py
@@ -2,7 +2,7 @@
 """
 # Public API
 from . import _widget as widget
-from ._functions import count, delete
+from ._api import count, delete
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -11,8 +11,8 @@ __license__ = 'MIT'
 
 def __init():
     import sys
-    from pytsite import assetman, odm, tpl, lang
-    from . import _model
+    from pytsite import assetman, odm, tpl, lang, events
+    from . import _model, _eh
 
     lang.register_package(__name__)
 
@@ -24,5 +24,7 @@ def __init():
     assetman.register_package(__name__)
     assetman.add('pytsite.flag@css/common.css', forever=True)
     assetman.add('pytsite.flag@js/common.js', forever=True)
+
+    events.listen('pytsite.odm.entity.delete', _eh.odm_entity_delete)
 
 __init()