+++ pytsite/flag/_eh.py
@@ -1,0 +1,12 @@
+"""Flag Package Event Handlers.
+"""
+from pytsite import odm as _odm
+
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+def odm_entity_delete(entity: _odm=_odm.Model):
+
+    pass