+++ pytsite/flag/_widget.py
@@ -1,6 +1,7 @@
 """Flag Package Widgets.
 """
 from pytsite import auth as _auth, widget as _widget, html as _html, tpl as _tpl, odm as _odm
+from . import _api
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -10,16 +11,16 @@ __license__ = 'MIT'
 class Flag(_widget.Base):
     """Flag Widget.
     """
-    def __init__(self, uid: str, **kwargs: dict):
+    def __init__(self, entity: _odm.Model, **kwargs: dict):
         """Init.
         """
         super().__init__(**kwargs)
-        self._uid = uid
+        self._entity = entity
         self._icon = kwargs.get('icon', 'fa fa-star')
 
     @property
-    def uid(self) -> str:
-        return self._uid
+    def entity(self) -> str:
+        return self._entity
 
     @property
     def icon(self) -> str:
@@ -27,12 +28,13 @@ class Flag(_widget.Base):
 
     @property
     def count(self) -> int:
-        return _odm.find('flag').where('uid', '=', self._uid).count()
+        return _api.count(self._entity)
 
     @property
     def flagged(self) -> bool:
-        f = _odm.find('flag').where('uid', '=', self._uid).where('author', '=', _auth.get_current_user())
-        return bool(f.count())
+        """Is entity flagged by current user.
+        """
+        return bool(_api.count(self._entity, _auth.get_current_user()))
 
     def render(self) -> _html.Element:
         current_user = _auth.get_current_user()
@@ -44,4 +46,4 @@ class Flag(_widget.Base):
         return _html.Span(_tpl.render('pytsite.flag@widget', {
             'widget': self,
             'current_user': current_user
-        }), cls=css, data_uid=self._uid)
+        }), cls=css, data_entity='{}:{}'.format(self._entity.model, self._entity.id))