+++ pytsite/flag/ep.py
@@ -1,7 +1,7 @@
 """Flag Package Endpoints.
 """
 from pytsite import auth as _auth, odm as _odm
-from . import _functions
+from . import _api
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -10,18 +10,15 @@ __license__ = 'MIT'
 
 def toggle(args: dict, inp: dict) -> dict:
     current_user = _auth.get_current_user()
-    uid = inp.get('uid')
+    entity = _odm.get_by_ref(inp.get('entity'))
 
-    print('TOGGLE')
+    if current_user.is_anonymous or not entity:
+        raise ValueError('Invalid input arguments.')
 
-    if current_user.is_anonymous or not uid:
-        return
+    if _api.count(entity, current_user):
+        _api.delete(entity, current_user)
+        return {'status': 'unflagged', 'count': _api.count(entity)}
+    else:
+        _api.flag(entity, current_user)
 
-    e = _odm.find('flag').where('uid', '=', uid).where('author', '=', current_user).first()
-    if e:
-        e.delete()
-        return {'status': 'unflagged', 'count': _functions.count(uid)}
-
-    _odm.dispense('flag').f_set('uid', uid).f_set('author', current_user).save()
-
-    return {'status': 'flagged', 'count': _functions.count(uid)}
+    return {'status': 'flagged', 'count': _api.count(entity)}