+++ pytsite/odm_ui/__init__.py
@@ -47,6 +47,6 @@ def __init():
     tpl.register_package(__name__)
     assetman.register_package(__name__)
 
-    events.listen('core.odm@register_model', _event_handlers.odm_register_model)
+    events.listen('pytsite.odm.register_model', _event_handlers.odm_register_model)
 
 __init()